Phone: Set MSISDN on boot from saved setting if SIM returned empty "My Phone Number"
Some carriers don't set the MSISDN (My Phone Number) and CM remedied
that by adding a setting in "Additional settings" in Call Settings
which saves a new MSISDN to the SIM. However, some phones fails to
save the MSISDN to the SIM and another remedy was added to set it
every boot from the stored value. However, the current implementation
only sets it if the framework returns a null value. This patch also
sets it if the framework returns an empty string.

Patch Set 3:
- Fixed simstate checking as the intent extra is a string not an
  integer
- Removed airplane mode checking as simstate will be triggered anyway

Change-Id: If8875b841d2029be12f367ab6f309bbfb81fc330
Signed-off-by: Pawit Pornkitprasan <p.pawit@gmail.com>