Phone: Double check mode before changing, part 2
A modem might report world phone mode when it really is only GSM or
CDMA capable, making the system switch to an invalid mode which
will later break the telephony code.
This simple check enforces the world phone overlay which is already
in place and will only consider it a correct value in that case.

This complements a commit with the same title by applying the same
logic to another code path that uses similar checks.

Change-Id: Ib7332511a693d40f53a79ba98844c9117871228a
Signed-off-by: Emilio López <turl@tuxfamily.org>