Phone: Double check mode before changing
A modem might report world phone mode when it really is only GSM or
CDMA capable, making the system switch to an invalid mode which
will later break the telephony code.
This simple check enforces the world phone overlay which is already
in place and will only consider it a correct value in that case.

Change-Id: Iedbc529497b2779f3a12ff56ca1b028cc9434227
Signed-off-by: Emilio López <turl@tuxfamily.org>